// ui/src/generated/WorkspaceSnapshotData.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BindingSummary } from "./BindingSummary";

export type WorkspaceSnapshotData = { objects: Array<BindingSummary>, total_bindings: number, truncated: boolean, working_directory: string, r_version: string, library_paths: Array<string>, namespace_paths: Array<string>, library_usage_complete: boolean, };