// ui/src/generated/Invocation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CapabilityRef } from "./CapabilityRef";
import type { Precondition } from "./Precondition";
import type { JsonValue } from "./serde_json/JsonValue";

export type Invocation = { client_request_id: string, capability: CapabilityRef, arguments: JsonValue, preconditions: Array<Precondition>, };