// ui/src/generated/MediaReference.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { OperationId } from "./OperationId";

export type MediaReference = { operation_id: OperationId, sequence: number, mime_type: string, byte_size: number, sha256: string, display_id: string | null, };